requirements.txt
streamlit
groq
yfinance
pandas
numpy
plotly
python-dotenv
textblob
sentence-transformers
faiss-cpu
